🦺 Build with /W4 under MSVC
fixes #863

+++ src/app/app-common/DCSMagneticModel.cpp
@@ -2,7 +2,10 @@
 //
 // Copyright (c) 2025 Fred Emmott <fred@fredemmott.com>
 //
-// This program is open source; see the LICENSE file in the root of the OpenKneeboard repository.
+// This program is open source; see the LICENSE file in the root of the
+// OpenKneeboard repository.
+
+#include "OpenKneeboard/numeric_cast.hpp"
 
 #include <OpenKneeboard/DCSMagneticModel.hpp>
 
@@ -109,7 +112,7 @@ float DCSMagneticModel::GetMagneticVariation(
   MAGtype_GeoMagneticElements geoElements {};
   MAG_Geomag(ellipsoid, sphereCoord, geoCoord, timedModel.get(), &geoElements);
 
-  return geoElements.Decl;
+  return numeric_cast<float>(geoElements.Decl);
 }
 
 }// namespace OpenKneeboard
