+++ src/app/app-common/UILayers/include/OpenKneeboard/FlyoutMenuUILayer.hpp
@@ -2,7 +2,8 @@
 //
 // Copyright (c) 2025 Fred Emmott <fred@fredemmott.com>
 //
-// This program is open source; see the LICENSE file in the root of the OpenKneeboard repository.
+// This program is open source; see the LICENSE file in the root of the
+// OpenKneeboard repository.
 #pragma once
 
 #include <OpenKneeboard/CursorClickableRegions.hpp>
@@ -11,9 +12,9 @@
 #include <OpenKneeboard/IToolbarItem.hpp>
 #include <OpenKneeboard/IUILayer.hpp>
 
-#include <shims/winrt/base.h>
-
 #include <OpenKneeboard/audited_ptr.hpp>
+
+#include <shims/winrt/base.h>
 
 #include <memory>
 
@@ -91,7 +92,7 @@ class FlyoutMenuUILayer final
   };
 
   struct Menu {
-    uint32_t mMargin {};
+    float mMargin {};
     D2D1_RECT_F mRect {};
     std::shared_ptr<CursorClickableRegions<MenuItem>> mCursorImpl;
     std::vector<D2D1_RECT_F> mSeparatorRects;
